PRODENG-2716 go mod update (#503)

+++ pkg/kubeclient/msr.go
@@ -239,7 +239,7 @@ func (kc *KubeClient) MSRURL(ctx context.Context, name string, rc dynamic.Resour
 		}
 
 		for _, p := range svc.Spec.Ports {
-			if p.Port == int32(externalPort) {
+			if int64(p.Port) == externalPort {
 				port = strconv.Itoa(int(p.NodePort))
 				break
 			}
